fix: update cancellation
